scampr/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.2
pydantic>=2.5.3
pydantic-settings>=2.1.0
pydantic[email]>=2.5.3
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
beanie>=1.24.0
bcrypt>=4.0.1
python-dotenv>=1.0.0